separate new tests into respective file
