+++ src/dom/components/__tests__/ReactDOMTextarea-test.js
@@ -116,36 +116,3 @@ describe('ReactDOMTextarea', function() {
     expect(console.warn.argsForCall.length).toBe(2);
   });
 });
-
-describe('ReactDOMInput', function() {
-  var React;
-  var ReactTestUtils;
-
-  var renderTextarea;
-
-  beforeEach(function() {
-    React = require('React');
-    ReactTestUtils = require('ReactTestUtils');
-
-    renderTextarea = function(component) {
-      var stub = ReactTestUtils.renderIntoDocument(component);
-      var node = stub.getDOMNode();
-      return node;
-    };
-  });
-
-  it('should display `defaultValue` of number 0', function() {
-    var stub = <input type="text" defaultValue={0} />;
-    var node = renderTextarea(stub);
-
-    expect(node.value).toBe('0');
-  });
-
-  it('should display `value` of number 0', function() {
-    var stub = <input type="text" value={0} />;
-    var node = renderTextarea(stub);
-
-    expect(node.value).toBe('0');
-  });
-
-});
